tests/requirements.txt
pytest>=8.0
pytest-asyncio>=0.23
requests>=2.31
grpcio>=1.60
grpcio-tools>=1.60
Pillow>=10.0
numpy>=1.24
aiohttp>=3.9
